[Jetcaster] SpotlessApply
