+++ Jetcaster/mobile/src/main/java/com/example/jetcaster/ui/home/ThemeViewModel.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.example.jetcaster.ui.home
 
 import androidx.lifecycle.ViewModel
@@ -8,8 +24,8 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ThemeViewModel @Inject constructor(
-    private val themeManager: ThemeManager
-): ViewModel() {
+    private val themeManager: ThemeManager,
+) : ViewModel() {
 
     private var currentThemeIndex = 0
 
@@ -18,4 +34,4 @@ class ThemeViewModel @Inject constructor(
         themeManager.updateTheme(theme)
         currentThemeIndex = (++currentThemeIndex) % AppTheme.entries.size
     }
-} 
+}
